Data.Functor.Invariant

Description

Haskell98 invariant functors (also known as exponential functors).

For more information, see Edward Kmett's article "Rotten Bananas":

http://comonad.com/reader/2008/rotten-bananas/

Invariant

class Invariant f where

Any * -> * type parametric in the argument permits an instance of Invariant.

Instances should satisfy the following laws:

invmap id id = id
invmap f2 f2' . invmap f1 f1' = invmap (f2 . f1) (f1' . f2')

Minimal complete definition

Nothing

Methods

invmap :: (a -> b) -> (b -> a) -> f a -> f b

default invmap :: (Generic1 f, Invariant (Rep1 f)) => (a -> b) -> (b -> a) -> f a -> f b

invmapFunctor :: Functor f => (a -> b) -> (b -> a) -> f a -> f b

Every Functor is also an Invariant functor.

GHC.Generics

With GHC 7.2 or later, Invariant instances can be defined easily using GHC generics like so:

{-# LANGUAGE DeriveGeneric, FlexibleContexts #-}

import Data.Functor.Invariant
import GHC.Generics

data T f a = T (f a) deriving Generic1

instance Invariant f => Invariant (T f)

Be aware that generic Invariant instances cannot be derived for data types that have function arguments in which the last type parameter appears in a position other than the result type (e.g., data Fun a = Fun (a -> a)). For these, you can derive them using the Data.Functor.Invariant.TH module.

genericInvmap :: (Generic1 f, Invariant (Rep1 f)) => (a -> b) -> (b -> a) -> f a -> f b

A generic implementation of invmap.

newtype WrappedFunctor f a

Wrap a Functor to be used as a member of Invariant.

Constructors

WrapFunctor

Fields

unwrapFunctor :: f a

invmapContravariant :: Contravariant f => (a -> b) -> (b -> a) -> f a -> f b

Every Contravariant functor is also an Invariant functor.

invmapProfunctor :: Profunctor p => (a -> b) -> (b -> a) -> p a a -> p b b

A Profunctor with the same input and output types can be seen as an Invariant functor.

invmapArrow :: Arrow arr => (a -> b) -> (b -> a) -> arr a a -> arr b b

An Arrow with the same input and output types can be seen as an Invariant functor.

newtype WrappedContravariant f a

Wrap a Contravariant functor to be used as a member of Invariant.

Constructors

WrapContravariant

Fields

unwrapContravariant :: f a

newtype InvariantProfunctor p a

A Profunctor with the same input and output types can be seen as an Invariant functor.

Constructors

InvariantProfunctor (p a a)

newtype InvariantArrow c a

An Arrow with the same input and output types can be seen as an Invariant functor.

Constructors

InvariantArrow (c a a)

Invariant2

class Invariant2 f where

Any * -> * -> * type parametric in both arguments permits an instance of Invariant2.

Instances should satisfy the following laws:

invmap2 id id id id = id
invmap2 f2 f2' g2 g2' . invmap2 f1 f1' g1 g1' =
  invmap2 (f2 . f1) (f1' . f2') (g2 . g1) (g1' . g2')

Methods

invmap2 :: (a -> c) -> (c -> a) -> (b -> d) -> (d -> b) -> f a b -> f c d

invmap2Bifunctor :: Bifunctor f => (a -> c) -> (c -> a) -> (b -> d) -> (d -> b) -> f a b -> f c d

Every Bifunctor is also an Invariant2 functor.

newtype WrappedBifunctor (p :: k -> k1 -> Type) (a :: k) (b :: k1)

Make a Functor over the second argument of a Bifunctor.

Constructors

WrapBifunctor

Fields

unwrapBifunctor :: p a b

invmap2Profunctor :: Profunctor f => (a -> c) -> (c -> a) -> (b -> d) -> (d -> b) -> f a b -> f c d

Every Profunctor is also an Invariant2 functor.

newtype WrappedProfunctor p a b

Wrap a Profunctor to be used as a member of Invariant2.

Constructors

WrapProfunctor

Fields

unwrapProfunctor :: p a b